Backend/requirements.txt
Flask
python-dotenv
Flask-SQLAlchemy
pymysql
Flask-JWT-Extended
marshmallow
cryptography
flask-cors
requests
gunicorn==21.2.0
joblib
pandas
lightgbm
pytest